move `consensus.cache_mutation_loci` to `consensus.mutation_extractor`
