+++ src/DAJIN2/core/consensus/__init__.py
@@ -2,4 +2,5 @@ from DAJIN2.core.consensus.consensus import call_consensus
 from DAJIN2.core.consensus.name_handler import call_allele_name
 from DAJIN2.core.consensus.name_handler import update_key_by_allele_name
 from DAJIN2.core.consensus.name_handler import add_key_by_allele_name
-from DAJIN2.core.consensus.clust_formatter import subset_clust, cache_mutation_loci
+from DAJIN2.core.consensus.clust_formatter import subset_clust
+from DAJIN2.core.consensus.mutation_extractor import cache_mutation_loci
